Cool hot fix

+++ ibs-db-services/src/main/java/ru/ibs/db/service/beans/GenreService.java
@@ -5,7 +5,7 @@ import ru.ibs.entity.Genre;
 import ru.ibs.repository.GenreRepository;
 
 @Service
-public class GenreService extends JpaService<Genre, Integer> {
+public class GenreService extends JpaService<Genre, Long> {
 
     private final GenreRepository repository;
